Make proto CallstackEvent a class
This gives us control over the interface (we can add the constructor) and the
implementation (e.g., 40 to 24 bytes, which will be useful to offset the
upcoming changes...).

This and many other protos in `capture_data.proto` are protos because they were
used in the old capture format, but that's gone now.

Bug: http://b/216089402

Test: Build, take some captures.