Replace the old timeline with the new one
This PR simply includes the new timeline instead of the old one. The
only difference is the scale, the old timeline always include 9 labels
and the new one shows between 2 and 10, depending on the zooming level.
There are several TODOs but they will be addressed soon.

TEST: Load a capture, test general actions, start/stop a capture.

Old Screenshots:
http://screen/4oJLcL6gaoY38Zf
http://screen/3RBQAMbeSmUFopr

New Screenshots:
http://screen/7SBa9jYhW4kVUzy
http://screen/4ynM7TkuwH67Web

Bug: http://b/208444071