Add ClientCaptureEventCollector::OnStopAndWaitRequestedByExternal
We would like to extract some linux capture functionalities from existing
`CaptureService` and `LinuxCaptureService`, and make them work both for
the native Orbit and the cloud collector.

Native Orbit works with the `GrpcClientCaptureEventCollector` which supports
calling stop and wait by external, while the cloud collector works with
the `UploaderClientCaptureEventCollector` which does not support external stop
as it will stop automatically after receiving and processing the
`CaptureFinishedEvent`.
The current `CaptureService` assumes a `GrpcClientCaptureEventCollector`
and calls `GrpcClientCaptureEventCollector::StopAndWait` in the
`CaptureService::FinalizeEventProcessing`.

With this change, we change
`GrpcClientCaptureEventCollector::StopAndWait` as a virtual method
`OnStopAndWaitRequestedByExternal` which is defined in
`ClientCaptureEventCollector` such that the `FinalizeEventProcessing`
method works for the `ClientCaptureEventCollector`.

Bug: http://b/216085824