Add "Select these callstacks" context menu action to CallTreeWidget (#3272)
This adds an action that allows to create a selection of callstack samples from
the top-down and bottom-up views. This allows to see where the samples that
correspond to a sub-tree are in the capture, and to scale the percentages of a
sub-tree in the selection. This also works on the existing selection.

We add `CallTreeNode::exclusive_callstack_events_`.
The new action collects them from under the selection in `CallTreeWidget` and
creates a new `CallstackEvent` selection from it.

A quick note that I had a brief look at whether this is particularly bad for
performance but nothing stood out.

Bug: http://b/168218963

Test: All kinds of selections on multiple captures.