Add FindModulePathAndBuildIdByAddress
This combines `GetModulePathByAddress` and `FindModuleBuildIdByAddress` in a
single function for efficiency, so that `FindModuleByAddress` only needs to be
called once.
`FindModuleBuildIdByAddress` is now unused and can be removed.

This speeds up `CreateTopDownViewFromPostProcessedSamplingData` and
`CreateBottomUpViewFromPostProcessedSamplingData` by 15 to 20% (as per the
timings we print in the logs with a capture with 4M samples).