Draw grey lines between labels of the the new Timeline
In this PR we are including MinorTicks, which are lines between labels
to give more precision to the user. As discussed with Carsten, they are
using the next scale to give users a sense of continuity when they are
zooming-in/out.

You can see more details in go/stadia-oribit-timeline-improvements (step
number 3).

http://screen/C6aYsVp8Z5KWbuP

Test: Unit Tests + Loading a capture + Start/Stop a capture.

Bug: http://b/208447247