Fix Layout properties in TrackContainer
We were overriten TimeGraphLayout *layout_ from CaptureViewElement
by a local variable layout_ which wasn't a pointer, therefore pointers
in TimeGraph and TrackPointer were different and we had the issue of
not being able to modify Imgui properties.

Also we are erasing the pointer in Track.h, since it's not used anymore.