gRPC-free DoCapture
This is what I had in mind regarding https://github.com/google/orbit/pull/3286
and its follow-ups. In my mind this succeeds in keeping a single "capture"
method on Linux.
Of course:
 - this might actually be worse, depending on the follow-up changes;
 - it might just not fit at all the plans for the cloud colletor.

Compared to the "Original Refactoring Plan" in
http://go/orbit-refactor-capture-service-for-cloud-collector, it avoids
`void*` by adding the two interfaces `ClientCaptureEventCollectorBuilder` and
`StartStopCaptureRequestWaiter`, and it avoids having to duplicate
`WaitForStopCaptureRequestOrMemoryThresholdExceeded`.

Of course gRPC `Capture` and gRPC-free `DoCapture` should be in appropriate
classes. I put everything in `CaptureService` and `LinuxCaptureService` just to
be faster.

Also this doesn't consider `WindowsCaptureService`.