Add GetCaptureTimeSpanNs to TimelineInfoInterface (#3295)
We are adding a method which returns the total number of nanoseconds
which the capture longs. This method will be used to print timestamps in
the new ISO-8601 format. Depending on the total size of the capture,
the displayed format will be different.

- We are also using this method instead of getting the number of microseconds
while possible because precision but the total refactor is not finished
(neither a priority).

Test: Load a capture, everything works.

http://b/170712621