Use ISO-8061 timestamps in the timeline
As a result of several discussions, we will introduce ISO-8601
timestamps behind a flag. After finishing the layout improvements in the
timeline, we will decide which timestamp we will finally use.

As it is explained in go/stadia-orbit-timeline-improvements, the iso
timestamps are generated using the timestamp, the total time of the
capture and the precision used (Same number of digits will be  used for
the labels).

For now, the flag is true, so we can test it easily.

Bug: http://b/162485497.

http://screen/3ScEoma8yqrJQSD, an example to show how it change
before/after a minute.