Replace DeferToBackgroundThreadAndWait
`ServiceDeployManager::Shutdown` uses `DeferToBackgroundThreadAndWait`
to synchronously execute functions on a different thread.

`DeferToBackgroundThreadAndWait` spins up an event loop to process
events while waiting.

That's not suitable since `Shutdown` is meant to be called from the
destructor. This has lead to crashing race conditions in the past.

The solution is to replace `DeferToBackgroundThreadAndWait` by a
blocking operation that does NOT processes events while waiting.

Bug: http://b/217645830