Handle chained info in PE/COFF unwinding in libunwindstack
Unwind info can be chained using the RUNTIME_FUNCTION member of the
UNWIND_INFO struct. This allows some optimizations like non-contiguous
code segments, or delaying saving registers in a function call. One
example are tail call optimizations where a function call is carried
out using a 'jmp' instruction and not a 'call' instruction. In all cases
the inner unwind infos (everything in the chain until the final chained
info) are considered "secondary", the final info in the chain represents
the "primary" unwind info (corresponding to an actual function call).

Info on chained unwind info can be found here:
https://docs.microsoft.com/en-us/cpp/build/exception-handling-x64?view=msvc-170#chained-unwind-info-structures
(There seems to be a typo with a mixup of volatile and nonvolatile
registers if I'm not mistaken.)

This change adds support for chained unwind info. We unit test chained
unwind info using a fake unwinder that omits the actual unwind info
parsing and caching, allowing a simple way to fake the chained info
(the alternative would be to write the unwind info to a fake object
file).

Tested: Unit tests.
Bug: http://b/194768602