Fix missing leftmost label in the Timeline (#3304)
Based on that labels are shown only when a tick is visible, if the left
major tick is not in screen, the label wasn't appearing. This confused
the user because we were losing the continuity while navigating through
the capture windows.

This is now fixed by including the previous major tick when rendering
labels.

http://screen/BaQgHujAqu7nens

First issue in http://b/217728115.