Solve both flickering and jumping timeline labels  (#3311)
* Solve both flickering and jumping timeline labels

In this PR we are solving 2 issues:

- Flickering: As labels were shown or not based on the size of the previous one,
there were some border cases which leads to skip some of them but not others.
http://screen/65dRc2myhGevgTy

- Jumping issue: As Orbit started by showing the left-most and was greedily
showing the next one which didn't intercept, the apparition of a new label in the
left part of the screen was swapping all the labels.
Before: http://screen/6H7uydk2dtxKEuW
After: http://screen/iuRBB8voXrMozme

To solve Flickering we now have the same number of skipped labels across all
visible ones. To solve jumping we take advantage of an invariant and used it to
always show the same set of labels given a time_range (the ones which are
multiple of the distance between them).

After making this, RenderLabels was a complete monster and so I split it into
several new methods which finally makes that method 10 lines shorter.

Test: Load a capture, make the capture windows as smaller as possible, play
with scales and also took a capture.

Ps: I know that this PR should have been at least 2, sorry for that. It should be
the last one before branch cut if I'm not missing something.

http://b/217728115

* Fixup: uint32_t + const