Add ToString.h and ORBIT_LOG_VAR/ORBIT_LOG_VARN macros
This is in essence reviving the PRINT_VAR/PRINT_VARN macros that were
once present in the Orbit codebase. The macros are now named
ORBIT_LOG_VAR/ORBIT_LOG_VARN. They are utility macros that log the name
and value of a variable in a simple invocation:

int my_int = 42;
ORBIT_LOG_VAR(my_int); // Logs "my_int = 42"
ORBIT_LOG_VARN(my_int, "new_name"); // Logs "new_name = 42"

Internally, the macro calls a new orbit_base::ToString(...) function
which is essentially a wrapper around std::to_string(...). The
difference is that we can add custom or not-supported types. Currently,
ToString adds support for string types, which is not supported by
std::to_string.

This is mainly useful for day-to-day debugging, and we don't expect too
many occurences to make their way to the official codebase. That being
said, there are cases where we actually want to log information in the
exact "name = value" syntax. We will rely on code reviews to push back
on misuse.

Tests: ToStringTest.cpp.