Fix function call highlighting for C-Functions
Some function names do not contain the parameter list (mostly C
functions). Thus, the disassembly view highlighter fails matching
those functions when occuring in a call.

This change restructures matching, such that not the complete match
will get highlighted, but the outermost capture group.
This allows more flexible describtion around the match without
highlighting this (similar to lookarounds, but without the restriction
of fixed size lookaheads).

With this change, we can now simplify the call target regex
to "anything that is inside parantheses and comes after a *call*".

Test: CodeViewerDemo + Disassembly of some functions
Bug: http://b/217849922