Make ETW timestamp format match that of QueryPerformanceCounter()
Use PROCESS_TRACE_MODE_RAW_TIMESTAMP when setting up the ETW trace so
that timestamps are not converted into system time but rather kept in
their original format, in our case, the format that matches
QueryPerformanceCounter timestamps.

Keep a copy of EVENT_TRACE_LOGFILE object returned by krabsetw so that
we can inspect its state and make sure the clock type is the one we
expect. A change logging this data structure will come in another PR but
we can already inspect the value of the log file structure in a debugger.

An issue was opened upstream in the krabsetw repo:
https://github.com/microsoft/krabsetw/issues/189

Test: locally tested, instrumented scopes now align with call stack
samples and scheduling events.