Allow user space instrumentation to instrument Windows functions
We want to be able to use user space dynamic instrumentation on targets running
under Wine. For this, we also have to consider the two Windows calling
conventions, "Microsoft x64" and "__vectorcall", when backing up registers in
`Trampoline.cpp`.
In particular, we have to back up the union of the registers used to pass
parameters (in the entry trampoline) and return values (in the exit trampoline)
in the three calling conventions, and back up registers that are volatile, i.e.,
caller-saved, on Linux but not on Windows (the Windows code can't expect the
Linux payload to possibly modify those registers).

Backing up more registers could lead to worse performance. As we know though,
the vast majority of the overhead is in `EntryPayload` and `ExitPayload`, which
are untouched. I tested this on a couple of simple benchmarks: the added
overhead seems to be in the range of measurement error, maybe ~20 ns per call in
the worst case by really wanting to see it.

Test: Tried on triangle.exe and a few "normal" Linux targets such as Trata.