Add MinHook to third_party
"The Minimalistic x86/x64 API Hooking Library for Windows"
This is a snapshot of https://github.com/TsudaKageyu/minhook at commit
4a455528f61b5a375b1f9d44e7d296d47f18bb18.

Note that we commented out MinHook's CMakeLists.txt "config" and "install"
sections to not interfere with our own packaging. This is the only change we 
made to the snapshot.