Append message on seccomp mode when SyscallInTracee fails because of it
We read the seccomp mode before executing the syscall and possibly report it in
the `ErrorMessage` if the syscall fails. We simply append a note; we might or
might not want to make it more visible in the future but that's out of scope
for now.

We also don't really want to call SyscallInTracee at all if the target is in
strict seccomp mode as we are sure we are going to crash it, but this will come
later.

NOTE: The two `ORBIT_FATAL`s in the code that restores registers and memory in
`SyscallInTracee` get changed to `ORBIT_ERROR`: Both fatal conditions were
triggered in the new test `SyscallInTraceeFailsBecauseOfSeccomp`.

NOTE: This change also disables seccomp altogether in Docker so that the new
tests can run.

Bug: http://b/219881622

Test:
 - Add unit test for `ReadSeccompModeOfThread`.
 - Add unit tests for the `ErrorMessage` of `MemoryInTracee::Create`.
 - Use user space instrumentation on pulseaudio. Use user space instrumentation
   on `triangle.exe` with the version of `GetExistingExecutableMemoryRegion`
   that starts from low addresses. Verify the error message.