ConnectToTargetDialog waits for the process (#3340)
The ConnectToTargetDialog now uses the ProcessManager to wait for the
process to be found, and only returns once this is the case. I deliverately
did not add a timeout because the user can cancel this at any time.

Bug: b/218502384
Test: Manual testing so far - works when the process is already running as
well as when the process is starting while the dialog is already open, and
tested aborting in multiple states. Will add a happy path E2E test next.