Add epilog detection and handling for PE/COFF in libunwindstack
Unwinding PE/COFF files uses unwind codes in the prolog and in the
body of a function. They contain information on the size of the prolog.
However, to detect whether one is in the body of a function or in the
epilog, one has to disassemble the machine code, and identify if it's an
epilog. Once idenfied, one unwinds accordingly: If in the body of a
function, one unwinds using all unwind codes. If in the epilog, one
unwinds by virtually executing the effect of the machine instructions
on the registers.

In this change, epilog detection and handling is implemented, following
the documentation on:
https://docs.microsoft.com/en-us/cpp/build/prolog-and-epilog?view=msvc-170#epilog-code

Tests: Unit tests.
Bug: http://b/194768602