Add argument list to symbol's name for PdbFileLlvm
Pdb's `ProcSym` debug record for function information does not
contain the argument types in their names. As a result, we can't
differentiate overloaded functions in Orbit.
However, the information is transitively available in the type
info section of the Pdb file.

This change attempts to retrieve the argument list from the
type info section and appends this to the function's name.

Note, this change only fixes the issue for the LLVM-based PDB
backend. A follow up change will target the DIA SDK implementation.
Also, there is a similar issue in the Coff file case, which will
be also addressed later.

Test: Load symbols for triangle.exe
Bug: http://b/219413222