Add no build id check to SymbolErrorDialog (#3357)
Add no build id check to SymbolErrorDialog

This adds a new check to SymbolErrorDialog, whether the module has a
build-id. If it doesn't, the "Add Symbol Location" Button is disabled,
because no matching with local symbol files can currently be done
without a build-id. Also changes the name of the dialog to "Symbol
Loading Error".

Test: Unit tests