Use relative positions when aligned a Track vertically (#3372)
An old issue that before didn't generate a bug just because Timegraph
position was (0,0). VerticallyMove into a Track need to use a relative
position since its aligned a Track into the respective container.

We are basically solving 2 issues, both moving a track and jumping to a
timer was aligned at (0,0) and not a TrackContainer position.

http://b/220682817

Test: Load a capture, see that moving tracks and jumping to timers works
as expected.