Synchronize the histogram with the function selected (#3371)
* Synchronize the histogram with the function selected

The histogram is updated only when the user clicks on one of the
live functions (on the upper-right). And the histogram is not
updated if the selection is removed or changed. Also, the
histogram is not updated when the selected function is changed
via clicking on a timer in the timeline.

Histogram update is prevented when the same function is selected.


Tests: Manual
Bug: http://b/221011110