Add DataView::GetActionVisibilities
With this change, we add the DataView::GetActionVisibilities method,
which returns a map with the visible action name as key, and action
availability as value.

The goal is to let the context menu of data views to have a consistent
behaviour with the top-down & bottom-up views: All actions that could be
invoked from this view are visible, and unavailable (i.e., discabled)
actions are greyed out.

Bug: http://b/205676296