Let DataView decide the menu order
With this change, we change the DataView to determine the menu order and
grouping in DataView::GetMenuWithGrouping, and change each individual
data views to decide action status (i.e., kInvisible,
kVisibleButDisabled, kVisibleAndEnabled) in GetActionStatus.

We termporaryly changed the unit tests which checks the context menu
entries to only check the appearances of visble and enabled actions (as
the previous tests).
The test will be refactored to also check the menu order and action
status.

Bug: http://b/205676296