Let DataView decide the menu order
With this change, we change the DataView to determine the menu order and
grouping in DataView::GetMenuWithGrouping, and change each individual
data views to decide which actions are visible and available in
GetActionVisibility.

We termporaryly changed the unit tests which checks the context menu
entries. The test will be refactored to also check the menu order.

Bug: http://b/205676296