Load debug symbols as raw struct instead of proto
This is done to have a better idea of the cost of loading symbols vs the
cost of creating protobufs. It opens the door to easy parallelization of
the expensive protobuf creation. Moreover, it centralizes this proto
creation step into a single function rather than having duplicate
implementations in all SymbolsFile implementations. Similarly, we defer
the expensive symbol demangling step to a bulk operation which could
also be easily parallelizable. Add ORBIT_SCOPE markers to measure the
run time of the symbol loading, demangling and proto creation steps.

Tests: compiles and passes unit tests. Profiling session.

https://screenshot.googleplex.com/7H83VZsAjeY2RKn