Optimize EntryPayload & ExitPayload by not using protos directly (#3390)
I had made the assumption that directly enqueuing
`orbit_grpc_protos::FunctionEntry`s and `orbit_grpc_protos::FunctionExit`s into
`LockFreeBufferCaptureEventProducer` would make sense as they only contain
primitive fields and are basically structs. This is true, but not when it comes
to their constructors and assignment operators.
We switch to using simple `struct`s instead. The conversion to protos is done in
`LockFreeBufferCaptureEventProducer::TranslateIntermediateEvent` (used in the
`LockFreeBufferCaptureEventProducer::ForwarderThread` thread), which is how this
class is intended to be used.

On a gamelet on the usual single threaded benchmark I use, that calls the
function I instrument one million times per second, the overhead on the
individual function call goes down from ~370 ns to ~290 ns.