Back up all registers that are volatile for Microsoft x64 in trampolines
We have observed some Windows functions (running on Wine) calling other
functions assuming that some registers are non-volatile even if they are
volatile according to the calling convention. This is likely an optimization
based on "Consider volatile registers destroyed on function calls unless
otherwise safety-provable by analysis such as whole program optimization" from
https://docs.microsoft.com/en-us/cpp/build/x64-calling-convention#callercallee-saved-registers.

We decide to just back up all registers that are volatile in the Microsoft x64
and __vectorcall conventions (except for x87 and MMX registers, it doesn't seem
to make a lot of sense), in addition to the ones we were already backing up.

This effectively adds:
- entry trampoline: R11, upper part of YMM8-15 if AVX;
- return trampoline: RCX, R8, R9, R10, R11, X/YMM4-5, upper part of YMM6-15 if
  AVX.

As RCX is now backed up, the return address of the `ExitPayload` is stored on
the stack and jumped to with a `ret`.

Some comments are re-worked.

Performance implications: on my usual single-threaded benchmark with the
instrumented function called one million times per second, the overhead is
measurable but doesn't exceed 20 ns.

Bug: http://b/219015201

Test: Tried on mentioned single-threaded benchmark, on Trata, on Windows binary
that showed the issue described above.