Make the margins around the histogarm constant
Previously the margins whould be 10 percent of width/height,
now they are constant and equal to the values suggested in
UX design. Also, in accordance with the design the labels
on the vertical exis report percents, not fractions.
The length of the ticks is also agjusted, as well as the
gap between them and the labels. The line width is doubled.

Tests: Manual
Bug: http://b/223351714