Add argument list to symbol's name using DIA (#3396)
This is the counter-part to #3354. Pdb's name fields are not
mangled and do not contain parameter type information for
functions. However, due to overloading and consistency, we want
to be able to show those to the user.

This change reads the parameter information using the `function
type` record and the Dia SDK.

Further, it extended the existing test for PDB parsing quite a
bit.

Test: Run on a couple of test pdbs and run the tests.
Bug: http://b/219413222