Make all histogram bars of the same size and avoid overlap (#3407)
Histograms bars overlap. This became visible with the
alternating coloring. This fixes this.

Their width might be rather different due to rounding errors.
That is, it's possible to have one bar of width 1 and the other
bar of width 4. The max difference in pixel-width of the bars
rendered should not exceed one 1.
To this end we generate width of each bin beforehand.

Tests: Manual, Unit tests
Bug: http://b/223566472