Add TaskGroup.h and Chunk.h
TaskGroup is a simple class used to parallelize a group of tasks.
Add tasks to be run in parallel then call "Wait()" to wait for all tasks
to complete. Note that "Wait()" is called from the destructor if it
wasn't explicitly called. This class is not thread-safe. A TaskGroup
object should be owned and accessed by a single thread.

Chunk.h contains a utility function, "CreateChunksOfSize", which is
useful in the context of creating a TaskGroup. It basically returns a
vector of absl::Span of the specified size from a given std::vector.

Add tests for both TaskGroup and Chunk.