String encoding conversion utility
Add StringConversion.h which contains "Narrow" and "Widen" functions
to convert to and from a wide string and a Utf-8 string. The function
names were taken from the Utf-8 manifesto, http://utf8everywhere.org/.
On Windows, wchar_t is a 16-bit wide UTF-16 unicode character that we
convert to and from UTF-8. On Linux, wchar_t is a 32-bit wide UTF-32
unicode character that we convert to and from UTF-8. The different
wchar_t representations require to specify a conversion facet per
platform (PlatformWStringConversionFacet). "Narrow" is a Utf-16 to Utf-8
conversion on Windows and a Utf-32 to Utf-8 conversion on Linux.

Replace truncating conversions from wide-string to string by
"Narrow(...)" calls in Windows code to properly convert strings
retrived from the Windows API (Utf-16) to Utf-8.

Add "/utf-8" compiler flag on Windows to set the source and execution
character sets to Utf-8. This is needed to make sure the narrow string
litterals in the code are Utf-8 encoded, which is mainly useful for the
tests, but it's nice to be clear about the string encoding we use in the
code.

Add StringConversionTest.cpp to test the string conversions.