Extend PersistantStorageManager to symbol mappings (#3424)
Extend PersistantStorageManager to symbol mappings

This adds the methods LoadModuleSymbolFileMappings and
SaveModuleSymbolFileMappings to PersistentStorageManager and its
implementation. This will be used to save the module to symbol file
overrides. The argument type is a hash map from module path (as the
module identifier) to a local file path (symbol file).

Test: Unit tests
Part of http://b/202391472