Don't inject USI or API if any target thread is in strict seccomp mode
The target will get killed during injection if the main thread is in strict
seccomp mode, but it might also get killed if any other thread in strict mode
hits a dynamically instrumented function or a manually instrumented scope.
See comments in code and bug.

Bug: http://b/219882987

Test: Add unit test for user space instrumentation. Tested on custom target with
various combinations of threads in strict seccomp mode.