Previously, we did not show the reason unwind failed (indicated by the type) in the top-down/bottom-up views.
This change addresses this the following ways:
For the top-down view, new nodes with the error type gets added below the
function for which unwinding failed.

For the bottom-up view, new nodes with the error type gets added below the
[unwind errors] node. The thread nodes will be shown below the respective
error nodes.

This also includes tooltips with a more detailed description of the error types.
Those texts will also be used in the tooltip in the callstack thread bars.

http://screenshot/64xfWQd9hkiBKPe
http://screenshot/AKxMFfgChVjKtPH

Test: Run Tests & Capture OrbitService&StadiaAdventure and inspect the views

Bug: http://b/188756080
Bug: http://b/224974922