Fix bug handling UWOP_SET_FPREG unwind code
The implementation of handling the UWOP_SET_FPREG unwind code was
implemented "the wrong way around", where the operation was carried out
to mirror the effect of the machine instructions. However, for unwinding
one has to do the inverse operation, that is, in this case setting
the stack pointer to the value of the frame pointer register minus
the specified offset.

Note that the operation to set the frame pointer register is not the
same as the standard frame pointer register use (where the prologue
starts with 'push rbp' and 'mov rsp, rbp'). Setting the frame pointer
register usually occurs in the middle or at the end of the prologue and
can also set the frame pointer register to an address in the middle of
the callframe already set up.

An example prolog with a set frame pointer operation looks like:
push rbp
push rsi
push r12
sub rsp, 40
lea rbp, [rsp + 8]    <-- This is the "set frame pointer operation"

Tested: Adjusted and ran unit tests.
Bug: http://b/194768602