Add a label for the timestamp pointed by the mouse
In this PR we are adding a new label for the timestamp pointed by the
mouse. This label has 2 more digits of precision.

The label can be found in http://screenshot/3DBDt7K8CVgiwUr.

I made a few refactors:
- I simplify kLeftMargin and kRightMargin and kPixelMargin by only using
  kLabelPadding for the 4 dimensions.
- The label has to be drawn in Draw(), since it has to be updated in
  every frame, so we need to save the number of digits of places needed
  for the labels.

http://b/218312292

TODO: Timeline should be below the green line pointed by the mouse, as
well the green overlay (right click). Will come in a following PR.