Remove mangled name from SymbolInfo
"SymbolInfo" contained two fields related to the symbol name: "name" and "demangled_name".
The "name" field was sometimes used to store a symbol's mangled name, and sometimes used
to store the bare function name, with no argument or return type information. On top of having
a dual purpose, the "name" field wasn't really needed at all. The only place where we would use
it is in the case where the "demangled_name" was not available, which should never happen.

Similarly, remove the "name" field from "FunctionInfo" which we used to store the "SymbolInfo"
"name" field. This saves on proto size, but more importantly reduces ambiguity.