Choose the precision for the histogram ticks dymamically
Instead of using the hard-coded number of decimals we
choose the minimal number of decimals s.t. all the ticks
have different values. We have to limit the maximal number of
decimals due to space limitations.

Futher, the tick is now renderred at the position driven by the
rounded value.

After rounding some ticks may fall off to the right or to the left.
Therefore, the maximal number of tics has been increased by 1.

Test: Manual
Bug: http://b/224967376