Add globoal worker thread pool
Add ThreadPool::GetWorkerThreadPool() which returns a shared_ptr to a
global instance of a fixed-sized thread pool. This is done in order to
minimize the boilerplate code of having to pass in a ThreadPool object
to any object wanting to parallelize tasks. This is similar in spirit to
std::async being available from anywhere.

Having a global thread pool required adding a call to ShutdownAndWait
in ThreadPool's destructor since we don't know when the thread pool will
be dereferenced last. This also means that we don't have to initialize
nor destroy the global thread pool explicitly. It is created on demand
on the first call to GetWorkerThreadPool() and is destroyed on program
exit.

Having access to this global thread pool also simplifies the use of the
TaskGroup object, which now uses this thread pool by default if no
executor is specified.

b/225063353

Tests: FunctionsDataView filtering, which uses a TaskGroup works as
expected.