Fix incorrect over-the-edge selection
If a selection is active and the left edge of the histogram is
greater than the  absolute minimum of the dataset and if the user
selects again right-to-left and releases to the left from zero,
the new selection will start from the absolute minimum of the dataset.

This happens because `LocationToValue` returns 0 instead of `min_value`.

This fixes it.

Tests: Manual
Bug: http://b/225403525