Add PresetFile::is_loaded_ and set it after loading
With this change, we added `PresetFile::is_loaded_` to indicate the
preset file is loaded or not, as well as the setter and getter methods.
For loadable and partially loadable presets, once being loaded, their
`is_loaded_` are set as true.

This enables us to visually indicate whether a preset file is loaded
correctly in the ui.

Bug: http://b/187411031