Erase trivial accessible classes
It's used to be a pain creating new CaptureViewElements since we had to
implement its own Accessible class. We want to make Scrollbars a
CaptureViewElement without needing to create a special class for its
Accessibility.

In this PR we are making AccessibleCaptureViewElement not a virtual
class anymore and in the constructor we are giving the parameters that
the classes needed: "Name", "Role" and "State". As a result, we don't
need a bunch of similar classes anymore.

We still need to have a different implementation for Tracks(FakeTrackTab
and FakeTimerPane), TriangleToggle (State is not constant) and TimeGraph
(parent is not a CaptureViewElement).

http://b/225903736
http://b/216496191