Add default worker thread pool
Add ThreadPool::GetDefaultThreadPool() which returns a pointer to a
global instance of a fixed-sized thread pool. This is done in order to
minimize the boilerplate code of having to pass in a ThreadPool object
to any object wanting to parallelize tasks.

A call to "ThreadPool::InitializeDefaultThreadPool()" will create the default
thread pool. We can also override the default thread pool by calling
"ThreadPool::SetDefaultThreadPool(...)". Note that if neither of these
functions are called, the default thread pool will be created on the first
call to "ThreadPool::GetDefaultThreadPool()".

Fix the "ThreadPool" destruction sequence that required a manual call to
"ShutdownAndWait()", which is now called from the destructor.

Having access to this global thread pool simplifies the use of the "TaskGroup"
object, which now uses the default thread pool if no executor is specified.

b/225063353

Tests: FunctionsDataView filtering, which uses a TaskGroup works as
expected.