Remove InstrumentFunctionTest.CheckX87StackEmptyAfterCall (#3472)
This test causes troubles with `--cpu=haswell`. The reason is probably that the
compiler inserts a load onto the x87 stack after `DoSomething()` and before our
inline assembly. One way to fix it might be to correctly specify `st(0)` to
`st(7)` as clobbers.
Nonetheless, as this test is quite hacky due to the mix of code and inline
assembly it uses, and as it doesn't provide much value, let's just remove it so
that the problem doesn't come back in a different form.

Bug: http://b/226301903