Add mouse wheel handler for CaptureViewElement (#3466)
This is the first step to move handling of mouse events into CaptureViewElement
instead of handling them directly in CaptureWindow. Mouse wheel functionality
is now handled by TimeGraph. In the next PRs, the mouse wheel will behave
differently depending on the mouse position, and handling is then delegated
to TimelineUI if it is the element underneath the mouse.

The same pattern will be used to handle MouseEnter, MouseLeave, MouseMove,
MouseDown and MouseRelease in the future to support non-pixel perfect picking
(http://go/stadia-orbit-picking).

Bug: http://b/224764161
Test: CaptureWindow integration tests pass