Fix use of wrong size when initializing .debug_frame section
The current code incorrectly uses the file size of the .debug_frame
section, which is the actual size plus any padding, which can can
contain bogus data. The bad data may incorrectly be parsed as a frame
description entry (FDE), which can cause unwinding to fail. This is
fixed by using the correct in-memory size ("vmsize") to avoid parsing
beyond the end.

The PeCoffInterface test is adjusted to include padding in the
.debug_frame section with bad data that leads to an incorrect FDE
being parsed. We validate that, despite the bad data, there is no FDE
entry in the parsed data structure corresponding to the start address
of the bad FDE.

In an end-to-end integration, this changes leads to unwind errors
dropping from ~6-8% to < 1% (which is probably as low as it gets).

Tested: Unit tests.
Bug: http://b/194768602