Add a capture length column to ItemModel
With this change, we add a "Capture length" column to the load capture
widget. If the capture length is unknown (e.g., previous files that we
haven't touched yet), we show "--" to indicate that the capture length
information is still missing. Also we will add a tooltip for this later.

Bug: http://b/198562119