Speedup E2E tests in the connection window
At least for local execution, the e2e tests have become significantly
slower as the connection window grew more complext - I suspect this is
due to the large amount of projects and recent captures that I
accumulated on my machine.

This changes the instance connection script to use `find_control`
instead of the magic pywinauto lookup, which executes a lot faster.