Widen CallTreeWidget's "Thread / Function" column on row expansion
When expanding a row of the top-down or bottom-up view, the newly visible nodes
could be completely hidden if the "Thread / Function" column was too narrow.
We now possibly widen the column so that the arrow and the first few characters
of these nodes are visible.

Keeping performance in mind, we have to distinguish between the case of a single
row expanded manually, and the case of many rows expanded programmatically with
a context menu action or with the search functionality.

Bug: http://b/218449940

Test: Multiple manual tests expanding rows manually, with the context menu
actions, and with the search bar.