Create BatcherInterface
In this PR I'm creating BatcherInterface. Batcher will be divided between
this class and PrimitiveAssembler. BatcherInterface is the part that will
be overwritten by OpenGlBatcher to implement the functionality.

Discussed in go/stadia-orbit-testing-ui-elements, the interface contains:

- AddBox/Line/Triangle methods -> Inserting
- ResetElements() & GetNumElements()
- GetLayers() & DrawLayers() -> Drawing
- GetUserData() -> Mostly used for tooltips

And 2 trivial implementations:
- GetBatcherId()
- Push/PopTranslations() -> This is strange, I know.
a) we might be interested in using inheritance instead of composition here?
b) Should we instead just rename BatcherInterface?
c) Should we moved this to every implementation?
d) Should we leave this as it is? I'm leaning toward this one but I'm open
to opinions.

Next step is linking OpenGlBatcher to BatcherInterface and renaming Batcher
to PrimitiveAssembler (Merged it in my local branch: https://github.com/karupayun/orbit/commit/1d4d264fc5f1f90c033b62bc2d0b22f45750401f)