Make mouse line visible in the timeline
In this PR we are decreasing the z-position of the timeline, so it is
drawn below the mouse green line. In addition, we are decreasing the
size of the green overlay so it's not being drawn on top of the timeline
(Open bug: http://b/226401787) and modifying the minor tick to be fully
white but with alpha value as Carsten suggested.

http://screenshot/yxKUy2iCcJbTxFm

Bug: https://b/226402406